DROTH-1537 UI: Change router to accept linkIds for all validator assets
