+++ digiroad2-oracle/src/main/resources/db/migration/V178__alter_table_inaccurat_asset_add_linkId.sql
@@ -4,8 +4,4 @@ alter table inaccurate_asset MODIFY ASSET_ID NUMBER NULL;
 
 alter table inaccurate_asset
 add constraint
-   asset_info UNIQUE (ASSET_ID, ASSET_TYPE_ID, MUNICIPALITY_CODE, ADMINISTRATIVE_CLASS);
-
-alter table inaccurate_asset
-add constraint
-   link_info UNIQUE (LINK_ID, ASSET_TYPE_ID, MUNICIPALITY_CODE, ADMINISTRATIVE_CLASS);
+   assetId_linkId UNIQUE (ASSET_ID, LINK_ID, ASSET_TYPE_ID);
