+++ digiroad2-oracle/src/main/scala/fi/liikennevirasto/digiroad2/process/LengthLimitValidator.scala
@@ -5,7 +5,7 @@ import fi.liikennevirasto.digiroad2.dao.pointasset.PersistedTrafficSign
 import fi.liikennevirasto.digiroad2.linearasset.PersistedLinearAsset
 import fi.liikennevirasto.digiroad2.service.pointasset.TrafficSignType
 
-class LengthLimitValidator extends MassLimitationValidator {
+class LengthLimitValidator extends SevenRestrictionsLimitationValidator {
   override def assetTypeInfo: AssetTypeInfo =  LengthLimit
   override val allowedTrafficSign: Set[TrafficSignType] = Set(TrafficSignType.MaximumLength)
